Add files via upload
What You're Aiming For
In this exercise, you will expand on the basic calculator program by adding more advanced mathematical operations and error handling using object-oriented programming.
Instructions
1.	Create a class called "Calculator" that contains the following:
o	A dictionary attribute to store the available mathematical operations and their corresponding functions
o	A method called "init" that initializes the dictionary with the basic mathematical operations (+, -, *, /) and corresponding functions
o	A method called "add_operation" that takes in two arguments: the operation symbol and the corresponding function. This method should add the new operation and function to the dictionary.
o	A method called "calculate" that takes in three arguments: the first number, the operation symbol, and the second number. This method should use the dictionary to determine the appropriate function to perform the calculation. It should also include error handling to check if the operation symbol is valid and if the input values are numbers. If an error is encountered, the method should print an error message and raise an exception.
2.	Create separate functions for the advanced mathematical operations (exponentiation, square root, logarithm) and use the "add_operation" method to add them to the calculator's dictionary.
3.	In the main program, create an instance of the Calculator class, and use a while loop that allows the user to continue performing calculations until they choose to exit.
4.	Use the input() function to get input from the user for the numbers and operation symbol.Use the math library for advanced mathematical operations
5.	Use the isinstance() function to check if the input is a number.